pyproject.toml
[project]
name = "biblioteca-digital-logistica-publica"
version = "0.1.0"
description = "Biblioteca Digital de Logística Pública do Estado de São Paulo"
requires-python = ">=3.11"
dependencies = [
    "django>=5.1,<6.0",
    "psycopg2-binary>=2.9",
    "gunicorn>=22.0",
    "django-environ>=0.11",
    "openpyxl>=3.1",
    "whitenoise>=6.7",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "pytest>=8.0",
    "pytest-django>=4.9",
]
ai = [
    "openai>=1.50",
    "tenacity>=9.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "portal.settings"
python_files = ["test_*.py"]
